made usage of ORM more concise
